skills/html2doc/requirements.txt
pypandoc
python-docx
